+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
 - Enh #91: Update JS dependencies
 - Fix #59: @ becomes @@ when copying mentioning in wiki editor
 - Enh #30: Tabulator Support
+- Enh #38: Save menu collapse state in local storage
 
 
 1.1.3 (Unreleased)